Implemented Writable object for the TTY

+++ kernel/src/object.rs
@@ -13,5 +13,5 @@ pub trait Writable: Object {
 
 pub trait Readable: Object {
     /// Reads the content of [`self`] and write them to [`buffer`]
-    fn read(&self, buffer: &mut [u8]) -> ObjectResult<u64>;
+    fn read(&self, buffer: &mut [u8]) -> ObjectResult<usize>;
 }
